feat: react split pane
