+++ vscode-react/src/components/akMonacoEditorModal.component.tsx
@@ -2,9 +2,12 @@ import React from "react";
 import { Editor } from "@monaco-editor/react";
 import { AKModal } from "@react-components";
 
-export const AKMonacoEditorModal = ({ hideModal, content }: { hideModal: () => void; content?: string }) => (
-	<AKModal wrapperClasses={["!bg-transparent"]} classes={["bg-[#00000050]", "rounded-none"]}>
-		<div className="flex justify-end cursor-pointer text-white font-extrabold pt-8 text-xl" onClick={() => hideModal()}>
+export const AKMonacoEditorModal = ({ onCloseClicked, content }: { onCloseClicked: () => void; content?: string }) => (
+	<AKModal wrapperClasses={["!bg-transparent z-50"]} classes={["bg-[#00000050]", "rounded-none"]}>
+		<div
+			className="flex justify-end cursor-pointer text-white font-extrabold pt-8 text-xl"
+			onClick={() => onCloseClicked()}
+		>
 			X
 		</div>
 		<div className="m-auto">
